pyproject.toml
[tool.poetry]
name = "helpdesk"
version = "0.1.0"
description = ""
authors = ["m.romantsov <romancov_dzr@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"
django-widget-tweaks = "^1.5.0"
celery = "^5.4.0"
redis = "^5.2.1"
eventlet = "^0.38.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
drf-yasg = "^1.21.8"
isort = "^5.13.2"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
django-ckeditor-5 = "^0.2.15"
coverage = "^7.6.10"
flake8 = "^7.1.1"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}
python-telegram-bot = "^21.10"
django-select2 = "^8.2.3"
requests = "^2.32.3"
googletrans = "^4.0.2"
polib = "^1.2.0"
google-genai = "^1.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
